requirement.txt
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cssselect==1.2.0
et-xmlfile==1.1.0
h11==0.14.0
idna==3.7
jmespath==1.0.1
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parsel==1.9.1
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.1
requests==2.31.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
w3lib==2.1.2
webdriver-manager==4.0.1
wsproto==1.2.0
XlsxWriter==3.2.0
